requirements.txt
pandas>=0.24.0
numpy>=1.13.0
pytorch>=1.1.0
fastai>=1.0.30